+++ CHANGELOG.rst
@@ -22,15 +22,16 @@ Unreleased (next version 0.2.0)
 
 Changed
 '''''''
-
 - all algorithms now work on a list of output variables, not a single one
 
+Removed
+'''''''
+- class `othoz.paragraph.types.Graph`, as it does not add to the functionality of the framework
 
 0.1.0 - 09.10.2019
 ------------------
 
 Added
 '''''
-
 - repository skeleton
 - initial version of source code